send db backup status via whastapp
