+++ db/index.ts
@@ -3,6 +3,7 @@ import { RecordType } from "./entity/RecordType.js";
 import { UserType } from "./entity/UserType.js";
 import cron from "node-cron";
 import backup from "../services/db-backup.js";
+import {sendWhatsAppMsg_API} from "../services/whatsapp.js";
 
 async function insertDefaultData() {
   const defaultRecordTypes = [
@@ -50,8 +51,10 @@ const initialize = () => {
         try {
           await backup();
           console.log('Backup done');
+          await sendWhatsAppMsg_API(972566252561, 'Backup done');
         } catch (err) {
           console.error('Backup failed:', err);
+          await sendWhatsAppMsg_API(972566252561, 'Backup failed: ' + err);
         }
       });
       console.log("Cron job scheduled for backup.");
